[refactor] Extract interfaces into the contract directory for reusability.
