+++ src/lib/contracts/IExitIntent.ts
@@ -1,0 +1,11 @@
+export type ExitIntentListener = () => void;
+
+export interface IExitIntent {
+    attach(...listeners: ExitIntentListener[]): IExitIntent;
+
+    isAttached(): boolean;
+
+    detach(): IExitIntent;
+
+    addListener(...listeners: ExitIntentListener[]): void;
+}
