+++ src/lib/contracts/ILoader.ts
@@ -1,0 +1,5 @@
+export interface ILoader {
+    show(): ILoader;
+
+    hide(): ILoader;
+}
